apps/scraper/core/parser.py
import requests
from bs4 import BeautifulSoup


def fetch_url(url):
    """发送请求获取HTML内容"""
    try:
        response = requests.get(url, timeout=10)
        response.raise_for_status() # 检查请求是否成功
        response.encoding = response.apparent_encoding # 自动检测编码
        return response.text
    except requests.exceptions.RequestException as e:
        print(f"Error fetching {url}: {e}")
        return None

def parse_school_data(html):
    """
    解析HTML，提取所需的院校数据（需要根据目标网站结构具体实现）
    返回一个包含院校信息的字典列表
    """
    soup = BeautifulSoup(html, 'html.parser')
    data = []
    
    # --- 伪代码：需要根据实际网页结构填充 --- 
    # 找到包含学校列表的容器
    # school_list_container = soup.find('div', class_='school-list')
    # if not school_list_container:
    #     return data
    
    # 遍历每个学校条目
    # for item in school_list_container.find_all('div', class_='school-item'):
    #     school_name = item.find('h3').text.strip()
    #     # ... 提取其他信息 (招生人数, 分数线等) ...
        
    #     # 获取近三年数据 (可能需要访问详情页)
    #     detail_link = item.find('a')['href']
    #     detail_html = fetch_url(detail_link)
    #     if detail_html:
    #         yearly_data = parse_school_detail(detail_html)
    #         # ... 合并数据 ...
            
    #     data.append({
    #         'name': school_name,
    #         # ... 其他字段 ...
    #     })
    # --- 伪代码结束 --- 
    
    print("解析函数 parse_school_data 需要根据目标网站具体实现！")
    return data

def parse_school_detail(html):
    """解析学校详情页HTML，提取近三年数据（需要根据目标网站结构具体实现）"""
    soup = BeautifulSoup(html, 'html.parser')
    yearly_data = {
        # year: { 'score': score, 'admission_count': count }
    }
    # --- 伪代码：需要根据实际网页结构填充 --- 
    # 找到包含历年数据的表格或列表
    # data_table = soup.find('table', id='history-data')
    # if data_table:
    #     for row in data_table.find_all('tr')[1:]: # 跳过表头
    #         cols = row.find_all('td')
    #         year = int(cols[0].text.strip())
    #         score = int(cols[1].text.strip()) # 假设是总分
    #         count = int(cols[2].text.strip())
    #         yearly_data[year] = {'score': score, 'admission_count': count}
    # --- 伪代码结束 --- 
    print("解析函数 parse_school_detail 需要根据目标网站具体实现！")
    return yearly_data 