requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
Pillow==10.0
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.3
openpyxl==3.1.2
# psycopg2-binary==2.9.9 # 移除 PostgreSQL 驱动
mysqlclient==2.2.0  # 添加 MySQL 驱动 (请确保安装了系统依赖)
# 或者备选: PyMySQL==1.1.0 
python-dotenv==1.0.0 